Tests/LibWeb/Text/expected/title.txt

1: ""
2a: 0
2b: 1
2c: "This is a title!"
2d: "This is a title!"
3: ""
4a: 3
4b: ""
4c: ""
4d: ""
4e: 3
4f: "This is another title!"
4g: ""
4h: ""
